+++ test/variant_value_construct_cx.cpp
@@ -82,7 +82,7 @@ int main()
 
         STATIC_ASSERT( v.index() == 1 );
         STATIC_ASSERT( holds_alternative<float>(v) );
-        STATIC_ASSERT( get<1>(v) == 3.14f );
+        STATIC_ASSERT( get<1>(v) == (float)3.14f ); // see FLT_EVAL_METHOD
     }
 
     {
@@ -97,7 +97,7 @@ int main()
 
         STATIC_ASSERT( v.index() == 2 );
         STATIC_ASSERT( holds_alternative<float>(v) );
-        STATIC_ASSERT( get<2>(v) == 3.14f );
+        STATIC_ASSERT( get<2>(v) == (float)3.14f );
     }
 
     {
